--- src/main/java/by/photoshop/smelnik/pageObjects/baseSettings/CookieXPath.java
@@ -1,5 +1,0 @@
-package by.photoshop.smelnik.pageObjects.baseSettings;
-
-public class CookieXPath {
-    public static final String COOKIE_CLOSE_XPATH = "//button[text()='Отклонить']";
-}
