--- src/main/java/by/photoshop/smelnik/pageObjects/baseSettings/LoginObjectXpath.java
@@ -1,9 +1,0 @@
-package by.photoshop.smelnik.pageObjects.baseSettings;
-
-public class LoginObjectXpath {
-    public static final String LOGIN_XPATH = "//div/span/span[@class='ok-auth__info']";
-    public static final String INPUT_LOGIN_XPATH = "//input[@class='form-control form-group  require_fields required data-input-check -state-empty']";
-    public static final String INPUT_PASSWORD_XPATH = "//input[@class='form-control  require_fields required data-input-check -state-empty']";
-    public static final String BUTTON_ENTER_XPATH = "//div[@class='ok-form-row']/button[contains(text(), 'Войти')]";
-    public static final String CHECK_LOGIN_XPATH = "//span[text()='sergei-msg@tut.by']";
-}
