--- src/main/java/by/photoshop/smelnik/pageObjects/baseSettings/PopUpWindowXpath.java
@@ -1,5 +1,0 @@
-package by.photoshop.smelnik.pageObjects.baseSettings;
-
-public class PopUpWindowXpath {
-    public static final String CLOSE_POPUP_WINDOW = "//*[@id=\"jivo_close_button\"]/jdiv";
-}
