--- src/main/java/by/photoshop/smelnik/pageObjects/cart/CartXPath.java
@@ -1,8 +1,0 @@
-package by.photoshop.smelnik.pageObjects.cart;
-
-public class CartXPath {
-    public static final String ADD_TO_CART = "//button[@class='ok-product__add-shcart ok-btn -btn-shopping-cart ok-product__btn ok-product__btn2']/*[1]";
-    public static final String GO_TO_CART = "//button[@class='ok-shcart__btn ok-shcart__ico']";
-    public static final String CART_CHECK = "//div[@class='ok-order__text']/a";
-
-}
