--- src/main/java/by/photoshop/smelnik/pageObjects/dataForTest/LoginData.java
@@ -1,6 +1,0 @@
-package by.photoshop.smelnik.pageObjects.dataForTest;
-
-public class LoginData {
-    public static final String RIGHT_LOGIN = "sergei-msg@tut.by";
-    public static final String RIGHT_PASSWORD = "rapid2132";
-}
