--- src/main/java/by/photoshop/smelnik/pageObjects/searchTest/SearchQueries.java
@@ -1,6 +1,0 @@
-package by.photoshop.smelnik.pageObjects.searchTest;
-
-public class SearchQueries {
-    public static final String SEARCH_QUERY1 = "Зеркальный фотоаппарат Canon EOS 5D Mark IV BODY";
-    public static final String SEARCH_RESULT1 = "Зеркальный фотоаппарат Canon EOS 5D Mark IV BODY.";
-}
