--- src/main/java/by/photoshop/smelnik/pageObjects/searchTest/SearchQueriesXPath.java
@@ -1,9 +1,0 @@
-package by.photoshop.smelnik.pageObjects.searchTest;
-
-public class SearchQueriesXPath {
-    public static final String SEARCH_FIELD_BASE_PAGE = "//input[@id='findtext']";
-    public static final String SEARCH_BUTTON_START_BASE_PAGE = "//button/i";
-    public static final String SEARCH_RESULT1 = "//span[@itemprop='name' and text() = 'Зеркальный фотоаппарат Canon EOS 5D Mark IV BODY.']";
-
-
-}
