+++ src/main/java/by/photoshop/smelnik/ui/data/DelayAndLength.java
@@ -1,0 +1,9 @@
+package by.photoshop.smelnik.ui.data;
+
+public class DelayAndLength {
+    public static final int MIN_DURATION = 3;
+    public static final int DEFAULT_DURATION = 5;
+    public static final int MAX_DURATION = 8;
+    public static final int MIN_PASSWORD = 6;
+    public static final int MAX_PASSWORD = 10;
+}
