+++ src/main/java/by/photoshop/smelnik/ui/data/LoginData.java
@@ -1,0 +1,6 @@
+package by.photoshop.smelnik.ui.data;
+
+public class LoginData {
+    public static final String RIGHT_LOGIN = "sergei-msg@tut.by";
+    public static final String RIGHT_PASSWORD = "rapid2132";
+}
