+++ src/main/java/by/photoshop/smelnik/ui/pages/cookies/CookieXPath.java
@@ -1,0 +1,5 @@
+package by.photoshop.smelnik.ui.pages.cookies;
+
+public class CookieXPath {
+    public static final String COOKIE_CLOSE_XPATH = "//button[text()='Отклонить']";
+}
