+++ src/main/java/by/photoshop/smelnik/ui/pages/popup/PopUpWindowXpath.java
@@ -1,0 +1,5 @@
+package by.photoshop.smelnik.ui.pages.popup;
+
+public class PopUpWindowXpath {
+    public static final String CLOSE_POPUP_WINDOW = "//*[@id=\"jivo_close_button\"]/jdiv";
+}
