+++ src/main/java/by/photoshop/smelnik/ui/pages/search/SearchQueries.java
@@ -1,0 +1,6 @@
+package by.photoshop.smelnik.ui.pages.search;
+
+public class SearchQueries {
+    public static final String SEARCH_QUERY1 = "Зеркальный фотоаппарат Canon EOS 5D Mark IV BODY";
+    public static final String SEARCH_RESULT1 = "Зеркальный фотоаппарат Canon EOS 5D Mark IV BODY.";
+}
