+++ src/main/java/by/photoshop/smelnik/ui/utils/Waiter.java
@@ -1,0 +1,11 @@
+package by.photoshop.smelnik.ui.utils;
+
+public class Waiter {
+    public static void wait(int sec) {
+        try {
+            Thread.sleep(sec * 1000);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+}
