--- src/test/java/by/photoshop/smelnik/ui_tests/SearchTest.java
@@ -1,25 +1,0 @@
-package by.photoshop.smelnik.ui_tests;
-
-import by.photoshop.smelnik.pageObjects.baseSettings.CookiePage;
-import by.photoshop.smelnik.pageObjects.baseSettings.PopUpPage;
-import by.photoshop.smelnik.pageObjects.searchTest.SearchPage;
-import by.photoshop.smelnik.pageObjects.searchTest.SearchQueries;
-import by.photoshop.smelnik.utils.Waiter;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-public class SearchTest extends BaseTest {
-
-    @Test
-    public void testSearchGoods() {
-        SearchPage searchPage = new SearchPage();
-        CookiePage cookiePage = new CookiePage();
-        cookiePage.rejectCookieWindow();
-        PopUpPage popUpPage = new PopUpPage();
-        popUpPage.closePopUp();
-        searchPage.enterSearchQuery();
-        searchPage.submitSearchQuery();
-        Waiter.wait(4);
-        Assertions.assertEquals(SearchQueries.SEARCH_RESULT1, searchPage.findSearchResult1());
-    }
-}
